confirmation message added in delete order

+++ src/Pages/Dashboard/DashboardComponents/User/Orders.jsx
@@ -3,6 +3,7 @@ import { useAuthState } from "react-firebase-hooks/auth"
 import { useNavigate } from "react-router-dom"
 import Spinner from "../../../../Components/Spinner"
 import auth from "../../../../firebase.init"
+import Swal from "sweetalert2"
 
 const Orders = () => {
 	const [user] = useAuthState(auth)
@@ -25,16 +26,30 @@ const Orders = () => {
 	}
 
 	const handleDiscard = (id) => {
-		fetch("http://localhost:4000/booking/" + id, {
-			method: "DELETE",
+		Swal.fire({
+			title: "Are you sure?",
+			text: "You won't be able to revert this!",
+			icon: "warning",
+			showCancelButton: true,
+			confirmButtonColor: "#3085d6",
+			cancelButtonColor: "#d33",
+			confirmButtonText: "Yes, delete it!",
+		}).then((result) => {
+			if (result.isConfirmed) {
+				fetch("http://localhost:4000/booking/" + id, {
+					method: "DELETE",
+				})
+					.then((response) => response.json())
+					.then((data) => {
+						if (data.success) {
+							const remaining = orders.filter(
+								(order) => order._id !== id
+							)
+							setOrders(remaining)
+						}
+					})
+			}
 		})
-			.then((response) => response.json())
-			.then((data) => {
-				if (data.success) {
-					const remaining = orders.filter((order) => order._id !== id)
-					setOrders(remaining)
-				}
-			})
 	}
 
 	return (
